Create BattleField
My very first java code. The original assignment was to make a pseudocode for this game, and I was intrigued what it would actually take to make it. So I did. It's incredibly lengthy and everything is hardcoded like crazy, but I'm still proud of it. I didn't even know syntax when I started typing.